linux-gen: pktio: make dpdk use generic private data field
